# server/src/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is an API for car comments",
    "version": "1.0.0",
    "title": "Cars API",
    "contact": {
      "email": "razbarry9392@gmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Development"
    },
    {
      "url": "http://localhost:9000",
      "description": "Prod"
    }
  ],
  "basePath": "/api",
  "paths": {
    "/cars": {
      "get": {
        "tags": ["Cars"],
        "summary": "Get all the cars",
        "description": "Get all the cars",
        "produces": ["application/json"],
        "parameters": [],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/carsResponse"
              }
            }
          },
          "400": {
            "description": "Invalid status value",
            "schema": {
              "$ref": "#/definitions/InvalidResponse"
            }
          },
          "404": {
            "description": "Invalid route",
            "schema": {
              "$ref": "#/definitions/InvalidResponse"
            }
          }
        }
      },
      "post": {
        "tags": ["Cars"],
        "summary": "Save the car",
        "description": "Save the car",
        "produces": ["application/json"],
        "consumes": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "car object",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "car": {
                  "type": "object",
                  "$ref": "#/definitions/Car"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/carsResponse"
              }
            }
          },
          "400": {
            "description": "Invalid status value",
            "schema": {
              "$ref": "#/definitions/InvalidResponse"
            }
          },
          "404": {
            "description": "Invalid route",
            "schema": {
              "$ref": "#/definitions/InvalidResponse"
            }
          }
        }
      }
    },
    "/cars/{id}": {
      "put": {
        "tags": ["Cars"],
        "summary": "Update car",
        "description": "Update car",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "car id that needs to be deleted",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "description": "car object",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "car": {
                  "type": "object",
                  "$ref": "#/definitions/Car"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/carsResponse"
              }
            }
          },
          "400": {
            "description": "Invalid status value",
            "schema": {
              "$ref": "#/definitions/InvalidResponse"
            }
          },
          "404": {
            "description": "Invalid route",
            "schema": {
              "$ref": "#/definitions/InvalidResponse"
            }
          }
        }
      },
      "delete": {
        "tags": ["Cars"],
        "summary": "Delete car",
        "description": "Delete car",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "car id that needs to be deleted",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/carsResponse"
              }
            }
          },
          "400": {
            "description": "Invalid status value",
            "schema": {
              "$ref": "#/definitions/InvalidResponse"
            }
          },
          "404": {
            "description": "Invalid route",
            "schema": {
              "$ref": "#/definitions/InvalidResponse"
            }
          }
        }
      }
    },
    "/comments": {
      "get": {
        "tags": ["Comments"],
        "summary": "Get all the comments",
        "description": "Get all the comments",
        "produces": ["application/json"],
        "parameters": [],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/carsResponse"
              }
            }
          },
          "400": {
            "description": "Invalid status value",
            "schema": {
              "$ref": "#/definitions/InvalidResponse"
            }
          },
          "404": {
            "description": "Invalid route",
            "schema": {
              "$ref": "#/definitions/InvalidResponse"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "carsResponse": {
      "type": "object",
      "properties": {
        "_id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "detail": {
          "type": "string"
        },
        "year": {
          "type": "integer"
        },
        "type": {
          "type": "string",
          "enum": ["SUV", "4x4", "Familial"]
        },
        "price": {
          "type": "integer"
        },
        "image": {
          "type": "string"
        },
        "createdAt": {
          "type": "string"
        },
        "updatedAt": {
          "type": "string"
        }
      }
    },
    "Car": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "detail": {
          "type": "string"
        },
        "year": {
          "type": "integer"
        },
        "type": {
          "type": "string",
          "enum": ["SUV", "4x4", "Familial"]
        },
        "price": {
          "type": "string"
        },
        "image": {
          "type": "string"
        }
      }
    },
    "InvalidResponse": {
      "type": "object",
      "properties": {
        "statusCode": {
          "type": "string"
        },
        "message": {
          "type": "string"
        }
      }
    }
  }
}